flask-app/requirements.txt
flask
openai>=1.2.0
python-dotenv